+++ src/components/editor-page/editor-pane/editor-pane.tsx
@@ -4,22 +4,24 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-import { Editor, EditorChange, ScrollInfo } from 'codemirror'
-import React, { useCallback, useEffect, useRef, useState } from 'react'
+import { Editor, EditorChange } from 'codemirror'
+import React, { useCallback, useRef, useState } from 'react'
 import { Controlled as ControlledCodeMirror } from 'react-codemirror2'
 import { MaxLengthWarningModal } from '../editor-modals/max-length-warning-modal'
-import { ScrollProps, ScrollState } from '../synced-scroll/scroll-props'
+import { ScrollProps } from '../synced-scroll/scroll-props'
 import { allHinters, findWordAtCursor } from './autocompletion'
 import './editor-pane.scss'
 import { createStatusInfo, defaultState, StatusBar, StatusBarInfo } from './status-bar/status-bar'
 import { ToolBar } from './tool-bar/tool-bar'
-import { handleUpload } from './upload-handler'
-import { handleFilePaste, handleTablePaste, PasteEvent } from './tool-bar/utils/pasteHandlers'
 import { useApplicationState } from '../../../hooks/common/use-application-state'
 import './codemirror-imports'
 import { setNoteContent } from '../../../redux/note-details/methods'
 import { useNoteMarkdownContent } from '../../../hooks/common/use-note-markdown-content'
 import { useCodeMirrorOptions } from './hooks/use-code-mirror-options'
+import { useOnEditorPasteCallback } from './hooks/use-on-editor-paste-callback'
+import { useOnEditorFileDrop } from './hooks/use-on-editor-file-drop'
+import { useOnEditorScroll } from './hooks/use-on-editor-scroll'
+import { useApplyScrollState } from './hooks/use-apply-scroll-state'
 
 const onChange = (editor: Editor) => {
   for (const hinter of allHinters) {
@@ -36,77 +38,18 @@ const onChange = (editor: Editor) => {
   }
 }
 
-interface DropEvent {
-  pageX: number
-  pageY: number
-  dataTransfer: {
-    files: FileList
-    effectAllowed: string
-  } | null
-  preventDefault: () => void
-}
-
 export const EditorPane: React.FC<ScrollProps> = ({ scrollState, onScroll, onMakeScrollSource }) => {
   const markdownContent = useNoteMarkdownContent()
   const maxLength = useApplicationState((state) => state.config.maxDocumentLength)
-  const smartPasteEnabled = useApplicationState((state) => state.editorConfig.smartPaste)
   const [showMaxLengthWarning, setShowMaxLengthWarning] = useState(false)
   const maxLengthWarningAlreadyShown = useRef(false)
   const [editor, setEditor] = useState<Editor>()
   const [statusBarInfo, setStatusBarInfo] = useState<StatusBarInfo>(defaultState)
   const ligaturesEnabled = useApplicationState((state) => state.editorConfig.ligatures)
 
-  const lastScrollPosition = useRef<number>()
-  const [editorScroll, setEditorScroll] = useState<ScrollInfo>()
-  const onEditorScroll = useCallback((editor: Editor, data: ScrollInfo) => setEditorScroll(data), [])
-
-  const onPaste = useCallback(
-    (pasteEditor: Editor, event: PasteEvent) => {
-      if (!event || !event.clipboardData) {
-        return
-      }
-      if (smartPasteEnabled) {
-        const tableInserted = handleTablePaste(event, pasteEditor)
-        if (tableInserted) {
-          return
-        }
-      }
-      handleFilePaste(event, pasteEditor)
-    },
-    [smartPasteEnabled]
-  )
-
-  useEffect(() => {
-    if (!editor || !onScroll || !editorScroll) {
-      return
-    }
-    const line = editor.lineAtHeight(editorScroll.top, 'local')
-    const startYOfLine = editor.heightAtLine(line, 'local')
-    const lineInfo = editor.lineInfo(line)
-    if (lineInfo === null) {
-      return
-    }
-    const heightOfLine = (lineInfo.handle as { height: number }).height
-    const percentageRaw = Math.max(editorScroll.top - startYOfLine, 0) / heightOfLine
-    const percentage = Math.floor(percentageRaw * 100)
-
-    const newScrollState: ScrollState = { firstLineInView: line + 1, scrolledPercentage: percentage }
-    onScroll(newScrollState)
-  }, [editor, editorScroll, onScroll])
-
-  useEffect(() => {
-    if (!editor || !scrollState) {
-      return
-    }
-    const startYOfLine = editor.heightAtLine(scrollState.firstLineInView - 1, 'local')
-    const heightOfLine = (editor.lineInfo(scrollState.firstLineInView - 1).handle as { height: number }).height
-    const newPositionRaw = startYOfLine + (heightOfLine * scrollState.scrolledPercentage) / 100
-    const newPosition = Math.floor(newPositionRaw)
-    if (newPosition !== lastScrollPosition.current) {
-      lastScrollPosition.current = newPosition
-      editor.scrollTo(0, newPosition)
-    }
-  }, [editor, scrollState])
+  const onPaste = useOnEditorPasteCallback()
+  const onEditorScroll = useOnEditorScroll(onScroll)
+  useApplyScrollState(editor, scrollState)
 
   const onBeforeChange = useCallback(
     (editor: Editor, data: EditorChange, value: string) => {
@@ -137,26 +80,7 @@ export const EditorPane: React.FC<ScrollProps> = ({ scrollState, onScroll, onMak
     [maxLength]
   )
 
-  const onDrop = useCallback((dropEditor: Editor, event: DropEvent) => {
-    if (
-      event &&
-      dropEditor &&
-      event.pageX &&
-      event.pageY &&
-      event.dataTransfer &&
-      event.dataTransfer.files &&
-      event.dataTransfer.files.length >= 1
-    ) {
-      event.preventDefault()
-      const top: number = event.pageY
-      const left: number = event.pageX
-      const newCursor = dropEditor.coordsChar({ top, left }, 'page')
-      dropEditor.setCursor(newCursor)
-      const files: FileList = event.dataTransfer.files
-      handleUpload(files[0], dropEditor)
-    }
-  }, [])
-
+  const onDrop = useOnEditorFileDrop()
   const onMaxLengthHide = useCallback(() => setShowMaxLengthWarning(false), [])
   const codeMirrorOptions = useCodeMirrorOptions()
